+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   * `ignoreWarnings` is a confusing phrasing. Added the `warningsToIgnore`
     spelling which will be used preferentially. This was better than making the
     options inconsistent, or spelling the new option `ignoreFiles`.
+* Improved error messages when validation of a polymer.json object fails.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.5.0] - 2017-11-21